validacao http status code direto pela fastapi.status

+++ tests/test_security.py
@@ -15,9 +15,7 @@ def test_jwt():
 
 
 def test_jwt_invalid_token(client):
-    response = client.delete(
-        "/users/1", headers={"Authorization": "Bearer token-invalido"}
-    )
+    response = client.delete("/users/1", headers={"Authorization": "Bearer token-invalido"})
 
     assert response.status_code == HTTPStatus.UNAUTHORIZED
     assert response.json() == {"detail": "Could not validate credentials"}
