+++ tests/test_todo.py
@@ -50,9 +50,7 @@ def test_list_todos_should_return_5_todos(session, client, user, token):
     assert len(response.json()["todos"]) == expected_todos
 
 
-def test_list_todos_pagination_should_return_2_todos(
-    session, user, client, token
-):
+def test_list_todos_pagination_should_return_2_todos(session, user, client, token):
     expected_todos = 2
     session.bulk_save_objects(TodoFactory.create_batch(5, user_id=user.id))
     session.commit()
@@ -65,13 +63,9 @@ def test_list_todos_pagination_should_return_2_todos(
     assert len(response.json()["todos"]) == expected_todos
 
 
-def test_list_todos_filter_title_should_return_5_todos(
-    session, user, client, token
-):
+def test_list_todos_filter_title_should_return_5_todos(session, user, client, token):
     expected_todos = 5
-    session.bulk_save_objects(
-        TodoFactory.create_batch(5, user_id=user.id, title="Test todo 1")
-    )
+    session.bulk_save_objects(TodoFactory.create_batch(5, user_id=user.id, title="Test todo 1"))
     session.commit()
 
     response = client.get(
@@ -82,9 +76,7 @@ def test_list_todos_filter_title_should_return_5_todos(
     assert len(response.json()["todos"]) == expected_todos
 
 
-def test_list_todos_filter_description_should_return_5_todos(
-    session, user, client, token
-):
+def test_list_todos_filter_description_should_return_5_todos(session, user, client, token):
     expected_todos = 5
     session.bulk_save_objects(
         TodoFactory.create_batch(5, user_id=user.id, description="description")
@@ -99,13 +91,9 @@ def test_list_todos_filter_description_should_return_5_todos(
     assert len(response.json()["todos"]) == expected_todos
 
 
-def test_list_todos_filter_state_should_return_5_todos(
-    session, user, client, token
-):
+def test_list_todos_filter_state_should_return_5_todos(session, user, client, token):
     expected_todos = 5
-    session.bulk_save_objects(
-        TodoFactory.create_batch(5, user_id=user.id, state=TodoState.draft)
-    )
+    session.bulk_save_objects(TodoFactory.create_batch(5, user_id=user.id, state=TodoState.draft))
     session.commit()
 
     response = client.get(
@@ -116,9 +104,7 @@ def test_list_todos_filter_state_should_return_5_todos(
     assert len(response.json()["todos"]) == expected_todos
 
 
-def test_list_todos_filter_combined_should_return_5_todos(
-    session, user, client, token
-):
+def test_list_todos_filter_combined_should_return_5_todos(session, user, client, token):
     expected_todos = 5
     session.bulk_save_objects(
         TodoFactory.create_batch(
@@ -155,18 +141,14 @@ def test_delete_todo(session, client, user, token):
     session.add(todo)
     session.commit()
 
-    response = client.delete(
-        f"/todos/{todo.id}", headers={"Authorization": f"Bearer {token}"}
-    )
+    response = client.delete(f"/todos/{todo.id}", headers={"Authorization": f"Bearer {token}"})
 
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {"message": "Task has been deleted successfully."}
 
 
 def test_delete_todo_error(client, token):
-    response = client.delete(
-        f"/todos/{10}", headers={"Authorization": f"Bearer {token}"}
-    )
+    response = client.delete(f"/todos/{10}", headers={"Authorization": f"Bearer {token}"})
 
     assert response.status_code == HTTPStatus.NOT_FOUND
     assert response.json() == {"detail": "Task not found."}
